+++ schema/postgres/v1.10.0.sql
@@ -1,1 +1,2 @@
-ALTER TABLE "hit" ADD COLUMN "screen" varchar(20);
+ALTER TABLE "hit" ADD COLUMN "screen_class" varchar(20);
+ALTER TABLE "screen_stats" ADD COLUMN "class" varchar(20);