+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 * improved logging
 * `NewGeoDB` now takes a config as its parameter with optional logger for debugging
-* added screen classes to hits
+* added screen classes to hits and statistics
+* screen sizes with 0 width or height are no longer processed and stored
 
 ## 1.9.3